LayoutTests/media/video-playsinline-expected.txt

RUN(internals.settings.setInlineMediaPlaybackRequiresPlaysInlineAttribute(true))
RUN(video.playsinline = false)
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'true') OK
RUN(internals.settings.setInlineMediaPlaybackRequiresPlaysInlineAttribute(true))
RUN(video.playsinline = true)
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
RUN(internals.settings.setInlineMediaPlaybackRequiresPlaysInlineAttribute(false))
RUN(video.playsinline = false)
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
RUN(internals.settings.setInlineMediaPlaybackRequiresPlaysInlineAttribute(false))
RUN(video.playsinline = true)
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
END OF TEST
